+++ src/composite_components/Tile/TileActionBar.tsx
@@ -1,5 +1,5 @@
 /* ======================================================================== *
- * Copyright 2024 HCL America Inc.                                          *
+ * Copyright 2026 HCL America Inc.                                          *
  * Licensed under the Apache License, Version 2.0 (the "License");          *
  * you may not use this file except in compliance with the License.         *
  * You may obtain a copy of the License at                                  *
@@ -27,6 +27,7 @@ export interface ITileActionBarProps {
   menuSize?: string;
   disabled?: boolean;
   hasThumbnail?: boolean;
+  isTrash?: boolean;
 }
 
 const TileActionBar: React.FC<ITileActionBarProps> = (props: ITileActionBarProps) => {
@@ -74,6 +75,7 @@ const TileActionBar: React.FC<ITileActionBarProps> = (props: ITileActionBarProps
           overflowTooltip={overflowTooltip}
           menuSize={menuSize}
           disabled={disabled}
+          isTrash={props.isTrash}
         />
         )}
       </>
@@ -86,6 +88,7 @@ const TileActionBar: React.FC<ITileActionBarProps> = (props: ITileActionBarProps
         menuSize={menuSize}
         disabled={disabled}
         hasThumbnail={hasThumbnail}
+        isTrash={props.isTrash}
       />
       )}
     </>