+++ src/Select/SelectMultiple.stories.tsx
@@ -1,5 +1,5 @@
 /* ======================================================================== *
- * Copyright 2024, 2025 HCL America Inc.                                    *
+ * Copyright 2024, 2026 HCL America Inc.                                    *
  * Licensed under the Apache License, Version 2.0 (the "License");          *
  * you may not use this file except in compliance with the License.         *
  * You may obtain a copy of the License at                                  *
@@ -16,6 +16,7 @@
 import React from 'react';
 import { StoryFn, Meta } from '@storybook/react';
 import { userEvent, within } from '@storybook/testing-library';
+import CaretDownIcon from '@hcl-software/enchanted-icons/dist/carbon/es/caret--down';
 import InformationIcon from '@hcl-software/enchanted-icons/dist/carbon/es/information';
 
 import Select, { SelectChangeEvent } from './Select';
@@ -115,6 +116,11 @@ export default {
       control: false,
       description: 'https://mui.com/material-ui/api/select/',
     },
+    customIcon: {
+      description: 'This can be used to add a custom icon replacing the default information icon for helper text.',
+      options: ['None', 'CaretDownIcon', 'InformationIcon'],
+      control: { type: 'radio' },
+    },
   },
 } as Meta<typeof Select>;
 
@@ -130,6 +136,18 @@ const Template: StoryFn<typeof Select> = (args) => {
     setValues(typeof value === 'string' ? value.split(',') : value);
   };
 
+  let customIcon: React.ComponentType<React.SVGProps<SVGSVGElement>> | undefined;
+  switch (args.customIcon as unknown as string) {
+    case 'CaretDownIcon':
+      customIcon = CaretDownIcon;
+      break;
+    case 'InformationIcon':
+      customIcon = InformationIcon;
+      break;
+    default:
+      customIcon = undefined;
+  }
+
   return (
     <Select
       {...args}
@@ -141,6 +159,7 @@ const Template: StoryFn<typeof Select> = (args) => {
         }
         return (selected as string[]).join(', ');
       }}
+      customIcon={customIcon}
     >
       <MenuItem key="None" value="None" disabled>
         <em>{args.placeholder}</em>
@@ -234,13 +253,5 @@ export const ExampleMultipleSelectFullWidth = {
   args: {
     ...ExampleMultipleSelect.args,
     fullWidth: true,
-  },
-};
-
-export const ExampleMultipleSelectWithCustomIcon = {
-  render: Template,
-  args: {
-    ...ExampleMultipleSelect.args,
-    customIcon: InformationIcon,
   },
 };