+++ src/TextField/TextField.stories.tsx
@@ -1,5 +1,5 @@
 /* ======================================================================== *
- * Copyright 2024, 2025 HCL America Inc.                                    *
+ * Copyright 2024, 2026 HCL America Inc.                                    *
  * Licensed under the Apache License, Version 2.0 (the "License");          *
  * you may not use this file except in compliance with the License.         *
  * You may obtain a copy of the License at                                  *
@@ -156,11 +156,29 @@ export default {
       control: false,
       description: 'https://mui.com/material-ui/api/text-field/',
     },
+    customIcon: {
+      description: 'This can be used to add a custom icon replacing the default information icon for helper text.',
+      options: ['None', 'CaretDownIcon', 'InformationIcon'],
+      control: { type: 'radio' },
+    },
   },
 } as Meta<typeof TextField>;
 
 const Template: StoryFn<typeof TextField> = (args) => {
   const [value, setValue] = React.useState(args.value ? args.value : '');
+
+  let customIcon: React.ComponentType<React.SVGProps<SVGSVGElement>> | undefined;
+  switch (args.customIcon as unknown as string) {
+    case 'CaretDownIcon':
+      customIcon = CaretDownIcon;
+      break;
+    case 'InformationIcon':
+      customIcon = InformationIcon;
+      break;
+    default:
+      customIcon = undefined;
+  }
+
   return (
     <TextField
       {...args}
@@ -168,6 +186,7 @@ const Template: StoryFn<typeof TextField> = (args) => {
       onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
         setValue(event.target.value);
       }}
+      customIcon={customIcon}
     />
   );
 };
@@ -324,13 +343,5 @@ export const ExampleTextFieldWithUnitSelector = {
         />
       </Box>
     );
-  },
-};
-
-export const ExampleTextWithCustomIcon = {
-  render: Template,
-  args: {
-    ...ExampleTextField.args,
-    customIcon: InformationIcon,
   },
 };